requirements.txt
python-telegram-bot==20.3
SQLAlchemy==1.4.46
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.1
pytz==2023.3
bcrypt==4.0.1